requirements.txt
discord.py=1.7.3
discord-components=1.0.1
asyncio=3.4.3